+++ src/ldo/hospexProfile.shapeTypes.ts
@@ -1,6 +1,6 @@
 import { ShapeType } from '@ldo/ldo'
-import { hospexProfileContext } from './hospexProfile.context'
 import { hospexProfileSchema } from './hospexProfile.schema'
+import { hospexProfileContext } from './hospexProfile.context'
 import { HospexProfile } from './hospexProfile.typings'
 
 /**